refine the sleeper related helpers
